+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this library will be documented in this file:
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v3.3.1](https://github.com/NexusPHP/cs-config/compare/v3.3.0...v3.3.1) - 2021-08-31
+
+- Fix case where comment is exactly `//`
+
 ## [v3.3.0](https://github.com/NexusPHP/cs-config/compare/v3.2.2...v3.3.0) - 2021-08-31
 
 - Fix grammar usages
@@ -12,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove references to `liaison/cs-config`
 - Add NoCodeSeparatorCommentFixer
 - Add SpaceAfterCommentStartFixer
+- Add support for custom fixers
 - Remove `@internal` annotation from classes in public API
 - Bump to php-cs-fixer v3.1
 - Add FixerGenerator